+++ CHANGES.rst
@@ -1,11 +1,16 @@
 Changelog for Rapid Photo Downloader
 ====================================
 
-0.9.0b3 (2017-04-xx)
+0.9.0b3 (2017-04-15)
 --------------------
 
- - Fixed bug where NFS backup destination might not appear as writeable even
-   though it is.
+ - Fixed bug where a warning dialog window could be wrongly issued about a
+   backup destination not being writeable even though it though it is.
+
+ - Fixed bug where tracking of bytes downloaded could occasionally fail when
+   file copy errors were encountered.
+
+ - Improved logging of file copy problems.
 
  - Fixed some translation bugs. Thanks to Jose Luis Tirado for pointing them
    out.